Add code to process_client_logs.sh
This commit adds the code necessary for process_client_logs to function partially. There is a problem with the regex expression. It does not catch lines that only have one number for the date ex: Aug 1 and it also does not catch lines with numbers in the username ex: test10.

Co-authored-by: Dante Miller <mill8622@morris.umn.edu>