Work on implemting create_username_dist.sh
This commit works on the implementation of the create_username_dist.sh script. As of right now, the script concatenates all the failed_login_data.txt files, grabs all the usernames, runs a sort and uniq command that counts how many times each username appears, and
then passes that to an awk command which grabs the username and count and then prints that in the format of data.addRow(['username', count]);. The output generated by the awk command is put into a new file called output4.txt.

Co-authored-by: Dante Miller <mill8622@morris.umn.edu>