Finish implementing create_username_dist.sh
This commit finishes the implementation of the create_username_dist.sh script. This commit adds the code necessary so that the script can call wrap contents and generate the proper html output. We also added the code so that the script makes a tmp file to work in and then delete after.

Co-authored-by: Dante Miller <mill8622@morris.umn.edu>o