Continue implementing process_logs.sh
This commit contiues the implementation of process_logs.sh. We modified the script to make dir with the name of the tar file, extract those contents to that file, and then call process_client_logs.sh. Right now, the script works with our test directory, but eventually it will do all of its work in a scratch directory.

Co-authored-by: Dante Miller <mill8622@morris.umn.edu>