Work on implementation of process_logs.sh
This commit modifes process_logs.sh so that it uses a scratch directory and then calls the helper scripts: create_country, create_hours, ect. This script also modifes some of those helper scripts so that they are compatible with process_logs.sh. All of the helper scripts were writing their results to a specific directory rather than using the general directory that they were given as an argument.

Co-authored-by: Dante Miller <mill8622@morris.umn.edu>